requirements.txt
flask
flask_cors
sqlite3
requests
llama-cpp-python
gunicorn